+++ src/test/java/cart/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.common.Money;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
